backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Aliyun DashScope (LLM/Embedding/VL)
dashscope==1.14.1

# LangChain Ecosystem
langchain==0.1.4
langchain-community==0.0.16
langgraph==0.0.20
langchain-core==0.1.16

# Vector Database
faiss-cpu==1.7.4
numpy==1.26.3

# Data Processing
pydantic==2.5.3
pydantic-settings==2.1.0

# Document Processing
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0
python-pptx==0.6.23
pdf2image==1.17.0
Pillow==10.2.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Database (Mock mode only)
pymysql==1.1.0
sqlalchemy==2.0.25

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-Levenshtein==0.23.0
pypinyin==0.50.0
websockets==12.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
